requirement.txt
app==0.0.1
certifi==2019.3.9
chardet==3.0.4
charset-normalizer==2.1.0
greenlet==1.1.2
idna==2.8
parse==1.19.0
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
slack-bolt==1.14.0
slack-sdk==3.17.2
slackclient==1.3.1
SQLAlchemy==1.4.39
urllib3==1.24.3
websocket-client==0.47.0
